==> Notebooks/Large_Screen_Processing/1.Process_sgRNA.ipynb <==
import scipy
import statistics
import scanpy as sc
import numpy as np
import pandas as pd
import matplotlib as mpl
import collections
from collections import defaultdict
import matplotlib
from matplotlib import pyplot as plt
import scipy.io as io
%matplotlib inline

matplotlib.rcParams['pdf.fonttype'] = 42

from _util_updated import nested_dict
from _util_updated import turn_point
from _util_updated import load_data
from _util_updated import filter_umi

csv_DIR = ''
sg_csv = (pd.read_csv(csv_DIR + 'DAP3_NAMES.csv', delimiter=',', header = None))
sg_csv.columns = ('Seq', 'ID')

sgRNA_DIR = ''

DAA259_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA259_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA259_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA259_SG_RAW.index = sg_seq
for col_id in DAA259_SG_RAW.columns:
    col_name_adj = col_id + '-1'
    new_columns.append(col_name_adj)
DAA259_SG_RAW.columns = new_columns

DAA260_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA260_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA260_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA260_SG_RAW.index = sg_seq
for col_id in DAA260_SG_RAW.columns:
    col_name_adj = col_id + '-2'
    new_columns.append(col_name_adj)
DAA260_SG_RAW.columns = new_columns


DAA261_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA261_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA261_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA261_SG_RAW.index = sg_seq
for col_id in DAA261_SG_RAW.columns:
    col_name_adj = col_id + '-3'
    new_columns.append(col_name_adj)
DAA261_SG_RAW.columns = new_columns


DAA262_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA262_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA262_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA262_SG_RAW.index = sg_seq
for col_id in DAA262_SG_RAW.columns:
    col_name_adj = col_id + '-4'
    new_columns.append(col_name_adj)
DAA262_SG_RAW.columns = new_columns

DAA263_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA263_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA263_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA263_SG_RAW.index = sg_seq
for col_id in DAA263_SG_RAW.columns:
    col_name_adj = col_id + '-5'
    new_columns.append(col_name_adj)
DAA263_SG_RAW.columns = new_columns

DAA264_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA264_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA264_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA264_SG_RAW.index = sg_seq
for col_id in DAA264_SG_RAW.columns:
    col_name_adj = col_id + '-6'
    new_columns.append(col_name_adj)
DAA264_SG_RAW.columns = new_columns

DAA265_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA265_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA265_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA265_SG_RAW.index = sg_seq
for col_id in DAA265_SG_RAW.columns:
    col_name_adj = col_id + '-7'
    new_columns.append(col_name_adj)
DAA265_SG_RAW.columns = new_columns

DAA266_SG_RAW = pd.read_csv(sgRNA_DIR + 'DAA266_matrix_featurecount.csv.gz', compression='gzip', header=0,index_col = 0, sep=',', quotechar='"')
sg_seq = []
new_columns = []
for sg in DAA266_SG_RAW.index:
    sg_seq.append(sg.split('_')[2])
DAA266_SG_RAW.index = sg_seq
for col_id in DAA266_SG_RAW.columns:
    col_name_adj = col_id + '-8'
    new_columns.append(col_name_adj)
DAA266_SG_RAW.columns = new_columns

DAA259_df_adj,cutoffs = filter_umi(DAA259_SG_RAW)

print(np.mean(cutoffs))
### Replace sgRNA Sequence with Name
temp_DAA259 = DAA259_df_adj
new_name = []
for seq in temp_DAA259.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA259.index = new_name      
DAA259_df_name = temp_DAA259.T

DAA260_df_adj,cutoffs = filter_umi(DAA260_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA260 = DAA260_df_adj
new_name = []
for seq in temp_DAA260.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA260.index = new_name    
DAA260_df_name = temp_DAA260.T

DAA261_df_adj,cutoffs = filter_umi(DAA261_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA261 = DAA261_df_adj
new_name = []
for seq in temp_DAA261.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA261.index = new_name    
DAA261_df_name = temp_DAA261.T

DAA262_df_adj,cutoffs = filter_umi(DAA262_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA262 = DAA262_df_adj
new_name = []
for seq in temp_DAA262.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA262.index = new_name    
DAA262_df_name = temp_DAA262.T

DAA263_df_adj,cutoffs = filter_umi(DAA263_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA263 = DAA263_df_adj
new_name = []
for seq in temp_DAA263.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA263.index = new_name    
DAA263_df_name = temp_DAA263.T

DAA264_df_adj,cutoffs = filter_umi(DAA264_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA264 = DAA264_df_adj
new_name = []
for seq in temp_DAA264.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA264.index = new_name    
DAA264_df_name = temp_DAA264.T

DAA265_df_adj,cutoffs = filter_umi(DAA265_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA265 = DAA265_df_adj
new_name = []
for seq in temp_DAA265.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA265.index = new_name    
DAA265_df_name = temp_DAA265.T

DAA266_df_adj,cutoffs = filter_umi(DAA266_SG_RAW)
print(np.mean(cutoffs))

### Replace sgRNA Sequence with Name
temp_DAA266 = DAA266_df_adj
new_name = []
for seq in temp_DAA266.index:
        name_index = sg_csv[sg_csv.Seq == seq].index
        sgRNA_name = sg_csv.ID[name_index].tolist()[0]
        new_name.append(sgRNA_name)
temp_DAA266.index = new_name    
DAA266_df_name = temp_DAA266.T

### Combine Matrix

All_df_adj = pd.concat([DAA259_df_name, DAA260_df_name, DAA261_df_name, DAA262_df_name, DAA263_df_name, DAA264_df_name, DAA265_df_name, DAA266_df_name]).fillna(value=0)

DAA259_df_adj,cutoffs = filter_umi(DAA259_SG_RAW)
DAA260_df_adj,cutoffs = filter_umi(DAA260_SG_RAW)
DAA261_df_adj,cutoffs = filter_umi(DAA261_SG_RAW)
DAA262_df_adj,cutoffs = filter_umi(DAA262_SG_RAW)
DAA263_df_adj,cutoffs = filter_umi(DAA263_SG_RAW)
DAA264_df_adj,cutoffs = filter_umi(DAA264_SG_RAW)
DAA265_df_adj,cutoffs = filter_umi(DAA265_SG_RAW)
DAA266_df_adj,cutoffs = filter_umi(DAA266_SG_RAW)

All_df_adj_seq = pd.concat([DAA259_df_adj.T, DAA260_df_adj.T, DAA261_df_adj.T, DAA262_df_adj.T, DAA263_df_adj.T, DAA264_df_adj.T, DAA265_df_adj.T, DAA266_df_adj.T]).fillna(value=0)

sg_cells = []

for cell in All_df_adj.index:
    if sum(All_df_adj.loc[cell]) > 0:
        sg_cells.append(cell)
All_df_adj_SG = All_df_adj.loc[sg_cells]

All_df_adj_SG.to_pickle('./DAP3_SGRNA_IND_072821.pkl')

All_df_adj_seq.to_pickle('./DAP3_SGRNA_IND_072821_seq.pkl')

### Grouped Name (Combine sgRNA for each target)

for column in All_df_adj_SG.columns:
    new_column = column.split(':')[0]
    All_df_adj_SG.rename(columns={column:new_column}, inplace = True)

All_df_adj_grouped = All_df_adj_SG.groupby(All_df_adj_SG.columns, axis=1).sum()

All_df_adj_grouped.to_pickle('./DAP3_SGRNA_GROUP_072821.pkl')
